+++ src/PlanManager.ts
@@ -206,7 +206,6 @@ const PlanManager = {
       };
     }
 
-    console.log(query);
     // TODO remove this eventually lol
     if (query === 'elena') {
       terms = ['comp sci 446'];