+++ src/app/ChangeLogPreview.ts
@@ -3,19 +3,19 @@ import { ChangeLogPreviewInformation } from '../types/BaseTypes';
 const clp: ChangeLogPreviewInformation = {
   items: [
     {
-      title: 'Course topics are finally here!',
+      title: 'Room finder links!',
+      description:
+        "You can now get a direct link to a room on Tech Room Finder right from Paper! No need to find the room name on Paper just to have to search it again on the room finder if you're not sure where in Tech your class is. Just open the info side card by clicking on a course section and click on the room name.",
+    },
+    {
+      title: 'UI and feature updates to the plan view.',
+      description:
+        'The plan view search UI has been updated and is now more similar to the search UI in the schedule view. You can also now remove the summer quarter from a year without having to remove all of the courses and refreshing the page.',
+    },
+    {
+      title: 'Course topics are finally here. (v2.3)',
       description:
         'Seminar courses (first-year seminars, special topics, special projects, etc.) and other courses with specified topics now have their topics displayed in Paper (schedule view). Find course topics in search results, on the schedule, and in the course info side card.',
-    },
-    {
-      title: 'Multiple meeting pattern support.',
-      description:
-        'Classes that have multiple meeting patterns (different meeting times on different days) are now supported on the schedule. Previously, only the first one was displayed, oops.',
-    },
-    {
-      title: 'Option to disable URL-based save data.',
-      description:
-        "By default, all of your plan and schedule data is stored in the URL. However, this can get annoying if your browser keeps routing you to an old plan or schedule link. If you don't really use the plan and schedule links, you can now disable this feature in the settings. Save data is still accessible from your account or your browser storage.",
     },
   ],
 };